Move apitester
