+++ internal/http/apitester.go
@@ -1,4 +1,4 @@
-package apitesting
+package http
 
 import (
 	"bytes"
